--- src/main/java/subway/domain/SubwayFarePolicy.java
@@ -1,6 +1,0 @@
-package subway.domain;
-
-public interface SubwayFarePolicy {
-
-    Money calculate(final Route route);
-}
