+++ src/main/java/subway/domain/policy/ChargePolicyComposite.java
@@ -1,7 +1,12 @@
-package subway.domain;
+package subway.domain.policy;
 
 import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Component;
+import subway.domain.Money;
+import subway.domain.policy.discount.DiscountCondition;
+import subway.domain.policy.discount.SubwayDiscountPolicy;
+import subway.domain.policy.fare.SubwayFarePolicy;
+import subway.domain.route.Route;
 
 import java.util.List;
