+++ src/main/java/subway/domain/policy/fare/SubwayFarePolicy.java
@@ -1,0 +1,9 @@
+package subway.domain.policy.fare;
+
+import subway.domain.Money;
+import subway.domain.route.Route;
+
+public interface SubwayFarePolicy {
+
+    Money calculate(final Route route);
+}
