+++ src/main/java/subway/domain/route/EdgeSection.java
@@ -1,4 +1,4 @@
-package subway.domain;
+package subway.domain.route;
 
 import org.jgrapht.graph.DefaultWeightedEdge;
