+++ src/main/java/subway/domain/section/Section.java
@@ -1,4 +1,7 @@
-package subway.domain;
+package subway.domain.section;
+
+import subway.domain.station.Station;
+import subway.domain.station.Stations;
 
 import java.util.Objects;
