+++ src/test/java/subway/domain/RouteTest.java
@@ -5,6 +5,12 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import subway.domain.line.Line;
+import subway.domain.route.EdgeSection;
+import subway.domain.route.Route;
+import subway.domain.section.Section;
+import subway.domain.station.Station;
+import subway.domain.station.Stations;
 
 import java.util.List;
 import java.util.stream.Collectors;
