+++ src/test/java/subway/domain/StationTest.java
@@ -2,6 +2,7 @@ package subway.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import subway.domain.station.Station;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
