+++ src/test/java/subway/domain/policy/fare/DefaultPricePolicyTest.java
@@ -1,10 +1,16 @@
-package subway.domain;
+package subway.domain.policy.fare;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import subway.domain.Money;
+import subway.domain.line.Line;
+import subway.domain.route.Route;
+import subway.domain.section.Section;
+import subway.domain.station.Station;
+import subway.domain.station.Stations;
 
 import java.util.List;
 import java.util.stream.Stream;
